BREAKING COMMIT - stop separating native and external profiles
After talking to weavejester we decided to simplify my PR.
It can be achieved by treating project.clj insertions as part of
`data` concept we get in `profile-data` multimethod.

Now to make it work it just needs to use `external-profiles/profile-data`
and `external-profiles/profile-files` functions to look for a function
that would return valid lists of data and files.